+++ server/controllers/board/cmt.js
@@ -2,18 +2,18 @@ const { User, Post_comment, sequelize } = require("../../models");
 
 module.exports = {
   post: async (req, res) => {
-    const { accessToken, user_id, postId, comment } = req.body;
-    //여기에 accessToken 확인하는 과정 추가할 예정입니다
-    console.log(postId);
+    console.log(req.query);
+    const { accessToken, user_id, postId, comment } = req.query;
+
     const cocodusMember = await User.findOne({
       where: { id: user_id },
     });
     if (!cocodusMember) {
-      return res.status(403).send("not Authorized"); //id가 일치하지 않으므로 더이상 진행할 필요가 없습니다
+      return res.status(401).send("not Authorized"); //id가 일치하지 않으므로 더이상 진행할 필요가 없습니다
     }
 
     if (isNaN(Number(postId))) {
-      console.log("post 번호가 Number type이 아님"); //만약 postId가 숫자가 아닐 경우
+      console.log(`post 번호가 ${typeof postId} type 입니다`); //만약 postId가 숫자가 아닐 경우
       return res.status(400).send("Not found post id");
     }
 
@@ -27,10 +27,10 @@ module.exports = {
     res.status(201).end();
   },
   get: async (req, res) => {
-    let postId = Number(req.body.postId);
-    console.log(postId);
-    if (isNaN(Number(postId))) {
-      console.log("post 번호가 Number type이 아님"); //만약 postId가 숫자가 아닐 경우
+    let postId = Number(req.query.postId);
+
+    if (isNaN(postId)) {
+      console.log(`get요청에 포함된 postId가 ${typeof postId} type 입니다`); //만약 postId가 숫자가 아닐 경우
       return res.status(400).send("Not found post id");
     }
 
@@ -39,39 +39,33 @@ module.exports = {
     INNER JOIN Users ON Post_comments.user_id = Users.id
     WHERE Post_id = ${postId};`
     );
-    // await Post_comment.findAll({
-    //   include: [
-    //     {
-    //       model: User,
-    //       attributes: ["name"],
-    //     },
-    //   ],
-    //   where: { post_id: postId },
-    //   attributes: ["id", "comment"],
-    // });
 
-    if (comment[0].length === 0) {
+    const commentArray = [...comment[0]];
+    console.log(commentArray);
+    if (commentArray.length === 0) {
       console.log("댓글이 없습니다");
       return res.status(200).send("no comments in this post");
     }
 
-    return res.status(200).json(comment[0]);
+    return res.status(200).json(commentArray);
     // res.status(200).end();
   },
   patch: async (req, res) => {
-    const { accessToken, user_id, comment_id, comment } = req.body;
+    const { accessToken, user_id, comment_id, comment } = req.query;
     //여기에 accessToken 확인하는 과정 추가할 예정입니다
 
     const cocodusMember = await User.findOne({
       where: { id: user_id },
     });
     if (!cocodusMember) {
-      return res.status(403).send("not Authorized"); //id가 일치하지 않으므로 더이상 진행할 필요가 없습니다
+      return res.status(401).send("not Authorized"); //id가 일치하지 않으므로 더이상 진행할 필요가 없습니다
     }
 
     if (isNaN(Number(comment_id))) {
-      console.log("post 번호가 Number type이 아님"); //만약 postId가 숫자가 아닐 경우
-      return res.status(400).send("Not found post id");
+      console.log(
+        `get요청에 포함된 comment_Id가 ${typeof comment_id} type 입니다`
+      ); //만약 postId가 숫자가 아닐 경우
+      return res.status(400).send("Not found comment id");
     }
 
     // const newComment =
@@ -95,7 +89,7 @@ module.exports = {
     });
 
     if (!cocodusMember) {
-      return res.status(403).send("not Authorized"); //id가 일치하지 않으므로 더이상 진행할 필요가 없습니다
+      return res.status(401).send("not Authorized"); //id가 일치하지 않으므로 더이상 진행할 필요가 없습니다
     }
     // const deleteComment =
     await Post_comment.destroy({
