+++ server/controllers/board/index.js
@@ -1,9 +1,8 @@
 module.exports = {
   all: require("./all"),
-  list: require("./list"),
-  like: require("./like"),
-  recruiting: require("./recruiting"),
   cmt: require("./cmt"),
-  writing: require("./writing"),
+  like: require("./like"),
   mypost: require("./mypost"),
+  recruiting: require("./recruiting"),
+  writing: require("./writing"),
 };
